Renamed popup to displayComment file
